fix: resolve clippy pedantic warnings (93.6% reduction)
Systematically fixed clippy warnings with pedantic mode enabled:
- Fixed cast_sign_loss, cast_precision_loss, cast_possible_truncation
- Fixed trivially_copy_pass_by_ref, match_same_arms, manual_let_else
- Fixed assigning_clones using clone_from() for efficiency
- Added documentation for functions returning Result (missing_errors_doc)
- Fixed doc_markdown warnings for IndexedDB backticks

Progress: 220 warnings → 14 warnings (93.6% reduction)

Remaining warnings:
- missing_panics_doc (2)
- used_underscore_binding (2)
- type_complexity (2)
- implicit_hasher (2)
- too_many_arguments (4)

+++ src/components/graph_canvas/time_labels.rs
@@ -141,7 +141,7 @@ pub fn draw_hour_labels(
         {
             let day = i / 24;
             let hour_in_day = i % 24;
-            draw_hour_label_with_day(ctx, hour_in_day as usize, day, adjusted_x, dims.top_margin);
+            draw_hour_label_with_day(ctx, hour_in_day, day, adjusted_x, dims.top_margin);
         }
     }
 
@@ -197,7 +197,7 @@ pub fn draw_hour_labels(
 
 fn draw_hour_label_with_day(
     ctx: &CanvasRenderingContext2d,
-    hour: usize,
+    hour: i32,
     day: i32,
     x: f64,
     top: f64,
